Implmement coverage upload as callable workflow  (#5)
The setup of coverage upload is not trivial for new projects. Also, when
something is broken, we often need to fix it across the whole
organization. So using a central workflow will reduce maintenance costs.

Based on
https://github.com/pyapp-kit/workflows/blob/65abcc0b091a0d45dbe3c589266160cc9b62f8dc/.github/workflows/upload-coverage.yml

It is passing https://github.com/napari/napari/pull/8367 

I do not have an idea for testing in this repository.